feat: QuestionReport와 AnswerReport 분리
