+++ src/main/java/qp/official/qp/converter/AnswerReportConverter.java
@@ -1,0 +1,26 @@
+package qp.official.qp.converter;
+
+import qp.official.qp.domain.mapping.AnswerReport;
+import qp.official.qp.web.dto.AnswerReportRequestDTO;
+import qp.official.qp.web.dto.AnswerReportResponseDTO;
+
+public class AnswerReportConverter {
+
+    public static AnswerReportResponseDTO.AnswerReportReturnDTO toAnswerReportTestDTO(AnswerReport answerReport){
+        return AnswerReportResponseDTO.AnswerReportReturnDTO.builder()
+            .answerId(answerReport.getAnswer().getAnswerId())
+            .userId(answerReport.getUser().getUserId())
+            .answerReportId(answerReport.getAnswerReportId())
+            .content(answerReport.getContent())
+            .createdAt(answerReport.getCreatedAt())
+            .build();
+    }
+
+
+    public static AnswerReport toAnswerReport(AnswerReportRequestDTO.AnswerReportDTO request){
+        return null;
+    }
+
+
+
+}
