+++ src/main/java/qp/official/qp/web/dto/AnswerReportRequestDTO.java
@@ -1,0 +1,17 @@
+package qp.official.qp.web.dto;
+
+import lombok.Getter;
+
+public class AnswerReportRequestDTO {
+
+    @Getter
+    public static class AnswerReportDTO{
+        Long userId;
+        Long answerId;
+        String content;
+    }
+
+
+
+
+}
