+++ src/main/java/qp/official/qp/web/dto/QuestionReportRequestDTO.java
@@ -2,14 +2,7 @@ package qp.official.qp.web.dto;
 
 import lombok.Getter;
 
-public class ReportRequestDTO {
-
-    @Getter
-    public static class AnswerReportDTO{
-        Long userId;
-        Long answerId;
-        String content;
-    }
+public class QuestionReportRequestDTO {
 
     @Getter
     public static class QuestionReportDTO{
@@ -17,7 +10,6 @@ public class ReportRequestDTO {
         Long questionId;
         String content;
     }
-
 
 
 }
